+++ i18n/en.js
@@ -811,6 +811,8 @@ const en = {
         "surveysCompleted": "You've finished all available Connect surveys. We will reach out to you when there are new surveys and study activities to complete. Thank you for your contributions to the study!",
         "mainBodyDHQ3Header": "Diet History Questionnaire III (DHQ III)",
         "mainBodyDHQ3Description": "Questions about your typical diet in the past year. After clicking \"start\", the survey will open in a new browser tab or window and you will see a standard Federal data security warning message. Click \"Continue to Questionnaire\" to start the survey.",
+        "mainBodyReturnOfResults2025Header": "2025 Return of Results Preference Survey",
+        "mainBodyReturnOfResults2025Description": "Questions about the types of results and information you’re interested in getting back from Connect research."
     },
     "notifications": {
         "read": "Read",
@@ -1839,6 +1841,7 @@ const en = {
         "mod2024ConnectExperienceSurvey": "2024 Connect Experience Survey",
         "modCancerScreeningHistorySurvey": "Cancer Screening History Survey",
         "modDietHistoryQuestionnaireIIIDHQIII": "Diet History Questionnaire III (DHQ III)",
+        "modROIPreference2025": "2025 Return of Results Preference Survey",
         "sessionInactiveTitle": "Inactive",
         "sessionInactive": "You were inactive for 20 minutes, would you like to extend your session?<div class=\"modal-footer\"><button type=\"button\" title=\"Close\" class=\"btn btn-dark log-out-user\" data-bs-dismiss=\"modal\">Log Out</button><button type=\"button\" title=\"Continue\" class=\"btn btn-primary extend-user-session\" data-bs-dismiss=\"modal\">Continue</button></div>",
         "homeTitle": "My Connect - Home",
